Endret mappestruktur
Endret hele filstrukturen. Hver klasse har egen fil. Gjør det enklere å jobbe. Diverse variabler er lagt i egen fil. Samme med fetching av bilder, lyder og musikk.
Movement på enemy må jobbes med.
