+++ objectContainer.Py
@@ -1,0 +1,13 @@
+
+
+def init(): 
+    global players
+    players = []
+    global bullets
+    bullets = [[]]
+    global platforms
+    platforms = []
+    global portals
+    portals = []
+    global enemyBullets
+    enemyBullets = []
